requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
qdrant-client>=1.7.0
ollama>=0.1.0
pyyaml>=6.0
pydantic>=2.0.0
docling>=2.0.0
